# [generate] add faster `stop_strings` stopping criteria