… docs/source/en/internal/generation_utils.md