@MaxBourdon 2025-08-28
Thank you for your PR!
I added some comments on the implementation details, hope they could be useful 😄