@gante 2025-08-29
@MaxBourdon thank you for the feedback, corrected 😉 

(there was another edge case failing: the last token fully contains the stop string, but doesn't start with stop string characters)